+++ packages/std/src/math/decimal256.rs
@@ -7,7 +7,7 @@ use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign};
 use std::str::FromStr;
 use thiserror::Error;
 
-use crate::errors::StdError;
+use crate::errors::{CheckedMultiplyRatioError, StdError};
 use crate::{OverflowError, Uint512};
 
 use super::Fraction;
@@ -130,16 +130,30 @@ impl Decimal256 {
 
     /// Returns the ratio (numerator / denominator) as a Decimal256
     pub fn from_ratio(numerator: impl Into<Uint256>, denominator: impl Into<Uint256>) -> Self {
+        match Decimal256::checked_from_ratio(numerator, denominator) {
+            Ok(value) => value,
+            Err(CheckedMultiplyRatioError::DivideByZero) => {
+                panic!("Denominator must not be zero")
+            }
+            Err(CheckedMultiplyRatioError::Overflow) => panic!("Multiplication overflow"),
+        }
+    }
+
+    /// Returns the ratio (numerator / denominator) as a Decimal256
+    pub fn checked_from_ratio(
+        numerator: impl Into<Uint256>,
+        denominator: impl Into<Uint256>,
+    ) -> Result<Self, CheckedMultiplyRatioError> {
         let numerator: Uint256 = numerator.into();
         let denominator: Uint256 = denominator.into();
         if denominator.is_zero() {
-            panic!("Denominator must not be zero");
+            return Err(CheckedMultiplyRatioError::DivideByZero);
         }
 
-        Self(
+        Ok(Self(
             // numerator * DECIMAL_FRACTIONAL / denominator
-            numerator.multiply_ratio(Self::DECIMAL_FRACTIONAL, denominator),
-        )
+            numerator.checked_multiply_ratio(Self::DECIMAL_FRACTIONAL, denominator)?,
+        ))
     }
 
     pub const fn is_zero(&self) -> bool {
@@ -688,6 +702,19 @@ mod tests {
     #[should_panic(expected = "Denominator must not be zero")]
     fn decimal256_from_ratio_panics_for_zero_denominator() {
         Decimal256::from_ratio(1u128, 0u128);
+    }
+
+    #[test]
+    fn decimal256_checked_from_ratio_does_not_panic() {
+        assert_eq!(
+            Decimal256::checked_from_ratio(1u128, 0u128),
+            Err(CheckedMultiplyRatioError::DivideByZero)
+        );
+
+        assert_eq!(
+            Decimal256::checked_from_ratio(Uint256::MAX, 1u128),
+            Err(CheckedMultiplyRatioError::Overflow)
+        );
     }
 
     #[test]