+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 - cosmwasm-std: Implement `checked_multiply_ratio` for
   `Uint64`/`Uint128`/`Uint256`
+- cosmwasm-std: Implement `checked_from_ratio` for `Decimal`/`Decimal256`
 
 ### Changed